requirements.txt
google-genai
newsdataapi
numpy<2
pandas
pybind11>=2.12
python-dotenv
requests
scikit-learn
sentence-transformers
